tests/wpt/tests/html/canvas/offscreen/the-canvas-state/2d.state.saverestore.stack.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.state.saverestore.stack</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.state.saverestore.stack</h1>
<p class="desc">save()/restore() can be nested as a stack</p>


<script>
var t = async_test("save()/restore() can be nested as a stack");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.lineWidth = 1;
  ctx.save();
  ctx.lineWidth = 2;
  ctx.save();
  ctx.lineWidth = 3;
  _assertSame(ctx.lineWidth, 3, "ctx.lineWidth", "3");
  ctx.restore();
  _assertSame(ctx.lineWidth, 2, "ctx.lineWidth", "2");
  ctx.restore();
  _assertSame(ctx.lineWidth, 1, "ctx.lineWidth", "1");
  t.done();

});
</script>
